# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GNV OnBoard API",
    "version": "1.0.0",
    "description": "API REST versionnée : préfixe canonique `/api/v1`. Les mêmes routes sont montées aussi sous `/api/*` (alias rétrocompatibilité) — dans OpenAPI, seul le préfixe `/api/v1` est documenté.\n\n**Swagger UI** : `GET /api-docs` (dev par défaut ; prod si `SWAGGER_ENABLED=true`). **Spec JSON** : `GET /api-docs.json`.",
    "contact": {
      "name": "GNV Team"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement"
    },
    {
      "url": "https://api.gnv-onboard.com",
      "description": "Serveur de production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authToken"
      }
    },
    "schemas": {
...
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "example": "movie"
          },
          "poster": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "episodes": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "PaginatedMovies": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Movie"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/v1/movies": {
      "get": {
        "summary": "Liste paginée des films/séries actifs",
        "description": "Pagination par défaut page=1, limit=20 (max 100). Cache Redis pour requêtes sans en-tête Authorization.",
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMovies"
                }
              }
            }
          },
          "503": {
            "description": "Service temporairement indisponible (corps avec data vide)"
          }
        }
      }
    },
    "/api/v1/movies/{id}": {
      "get": {
        "summary": "Détail d'un film ou d'une série",
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          },
          "404": {
            "description": "Non trouvé"
          }
        }
      }
    },
...
  }
}